\name{dat.mtct.rob}
\alias{dat.mtct.rob}
\docType{data}
\title{
Example Dataset
}
\description{
from MTCT correlates study, C-section only
}
\usage{data("dat.mtct.rob")}
\format{
  A data frame with 55 observations on the following 2 variables.
  \describe{
    \item{\code{y}}{a numeric vector}
    \item{\code{V3_BioV3B_500}}{a numeric vector}
  }
}
\references{
Fong and Huang (2016) Modified Wilcoxon-Mann-Whitney Test and Power against Strong Null.
}
