\name{multinom.test}
\alias{multinom.test}
\title{
Multinom Test
}
\description{
Perform multinom test.
}
\usage{
multinom.test(X, Y, alternative = c("two.sided", "less", "greater"),
         correct = FALSE, perm = NULL, mc.rep = 10000, method =
         c("exact.2", "large.0", "large", "exact", "exact.0",
         "exact.1", "exact.3"), verbose = FALSE, mode =
         c("test", "var"), useC = TRUE)
}
\arguments{
  \item{X}{
X
}
  \item{Y}{
Y
}
  \item{alternative}{
alternative
}
  \item{correct}{
correct
}
  \item{perm}{
perm
}
  \item{mc.rep}{
mc.rep
}
  \item{method}{
method
}
  \item{verbose}{
verbose
}
  \item{mode}{
mode
}
  \item{useC}{
useC
}
}
